ref(metrics): Move MetricName into base-schema
