+++ relay-metrics/src/bucket.rs
@@ -2,7 +2,6 @@ use std::collections::{BTreeMap, BTreeSet};
 use std::hash::Hash;
 use std::iter::FusedIterator;
 use std::num::NonZeroU32;
-use std::sync::Arc;
 use std::{fmt, mem};
 
 use hash32::{FnvHasher, Hasher as _};
@@ -12,8 +11,8 @@ use serde::{Deserialize, Serialize};
 use smallvec::SmallVec;
 
 use crate::protocol::{
-    self, hash_set_value, CounterType, DistributionType, GaugeType, MetricResourceIdentifier,
-    MetricType, SetType,
+    self, hash_set_value, CounterType, DistributionType, GaugeType, MetricName,
+    MetricResourceIdentifier, MetricType, SetType,
 };
 use crate::{FiniteF64, MetricNamespace, ParseMetricError};
 
@@ -346,118 +345,6 @@ impl BucketValue {
         }
 
         Ok(())
-    }
-}
-
-/// Optimized string represenation of a metric name
-///
-/// The contained name does not need to be valid MRI, but it usually is.
-///
-/// The metric name can be efficiently cloned.
-#[derive(Debug, Clone, PartialEq, Eq, PartialOrd, Ord, Hash, Deserialize, Serialize)]
-#[serde(transparent)]
-pub struct MetricName(Arc<str>);
-
-impl MetricName {
-    /// Extracts the namespace from a well formed MRI.
-    ///
-    /// Returns [`MetricNamespace::Unsupported`] if the metric name is not a well formed MRI.
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// use relay_metrics::{MetricName, MetricNamespace};
-    ///
-    /// let name = MetricName::from("foo");
-    /// assert_eq!(name.namespace(), MetricNamespace::Unsupported);
-    /// let name = MetricName::from("c:custom_oops/foo@none");
-    /// assert_eq!(name.namespace(), MetricNamespace::Unsupported);
-    ///
-    /// let name = MetricName::from("c:custom/foo@none");
-    /// assert_eq!(name.namespace(), MetricNamespace::Custom);
-    /// ```
-    pub fn namespace(&self) -> MetricNamespace {
-        self.try_namespace().unwrap_or(MetricNamespace::Unsupported)
-    }
-
-    /// Extracts the namespace from a well formed MRI.
-    ///
-    /// If the contained metric name is not a well formed MRI this function returns `None`.
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// use relay_metrics::{MetricName, MetricNamespace};
-    ///
-    /// let name = MetricName::from("foo");
-    /// assert!(name.try_namespace().is_none());
-    /// let name = MetricName::from("c:custom_oops/foo@none");
-    /// assert!(name.try_namespace().is_none());
-    ///
-    /// let name = MetricName::from("c:custom/foo@none");
-    /// assert_eq!(name.try_namespace(), Some(MetricNamespace::Custom));
-    ///
-    /// ```
-    pub fn try_namespace(&self) -> Option<MetricNamespace> {
-        // A well formed MRI is always in the format `<type>:<namespace>/<name>[@<unit>]`,
-        // `<type>` is always a single ascii character.
-        //
-        // Skip the first two ascii characters and extract the namespace.
-        let maybe_namespace = self.0.get(2..)?.split('/').next()?;
-
-        MetricNamespace::all()
-            .into_iter()
-            .find(|namespace| maybe_namespace == namespace.as_str())
-    }
-}
-
-impl PartialEq<str> for MetricName {
-    fn eq(&self, other: &str) -> bool {
-        self.0.as_ref() == other
-    }
-}
-
-impl fmt::Display for MetricName {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        self.0.fmt(f)
-    }
-}
-
-impl From<String> for MetricName {
-    fn from(value: String) -> Self {
-        Self(value.into())
-    }
-}
-
-impl From<Arc<str>> for MetricName {
-    fn from(value: Arc<str>) -> Self {
-        Self(value)
-    }
-}
-
-impl From<&str> for MetricName {
-    fn from(value: &str) -> Self {
-        Self(value.into())
-    }
-}
-
-impl std::ops::Deref for MetricName {
-    type Target = str;
-
-    fn deref(&self) -> &Self::Target {
-        self.0.deref()
-    }
-}
-
-impl AsRef<str> for MetricName {
-    fn as_ref(&self) -> &str {
-        self.0.as_ref()
-    }
-}
-
-impl std::borrow::Borrow<str> for MetricName {
-    fn borrow(&self) -> &str {
-        self.0.borrow()
     }
 }
